refactor(feature:recent-transaction): initial setup and directory restructure
