+++ feature/recent-transaction/src/commonMain/composeResources/drawable/ic_error_black_24dp.xml
