+++ feature/recent-transaction/src/commonMain/composeResources/drawable/ic_local_atm_black_24dp.xml
